Move server-constants.js to app/
