+++ __tests__/components/document-head.test.tsx
@@ -1,9 +1,9 @@
-import React from 'react';
-import { createRoot } from 'react-dom/client';
+import React from 'react'
+import { createRoot } from 'react-dom/client'
 import DocumentHead from '../../components/document-head'
 
 const mockNextPublicURL = jest.fn()
-jest.mock('../../lib/notion/server-constants', () => ({
+jest.mock('../../app/server-constants', () => ({
   get NEXT_PUBLIC_URL() {
     return mockNextPublicURL()
   },
@@ -17,8 +17,8 @@ beforeEach(() => {
 })
 
 afterEach(() => {
-  document.body.removeChild(container);
-  container = null;
+  document.body.removeChild(container)
+  container = null
 })
 
 describe('DocumentHead', () => {
