+++ lib/gtag.js
@@ -1,4 +1,4 @@
-import { NEXT_PUBLIC_GA_TRACKING_ID } from './notion/server-constants'
+import { NEXT_PUBLIC_GA_TRACKING_ID } from '../app/server-constants'
 
 // https://developers.google.com/analytics/devguides/migration/measurement/virtual-pageviews
 export const pageview = (title, url) => {
