+++ lib/notion/blog-index-cache.js
@@ -7,22 +7,22 @@ const {
   NOTION_API_SECRET,
   DATABASE_ID,
   BLOG_INDEX_CACHE,
-} = require('./server-constants')
+} = require('../../app/server-constants')
 const { Client } = require('@notionhq/client')
 
 const notionClient = new Client({
   auth: NOTION_API_SECRET,
 })
 
-exports.exists = function() {
+exports.exists = function () {
   return fs.existsSync(BLOG_INDEX_CACHE)
 }
 
-exports.get = function() {
+exports.get = function () {
   return JSON.parse(fs.readFileSync(BLOG_INDEX_CACHE))
 }
 
-exports.set = async function() {
+exports.set = async function () {
   let params = {
     database_id: DATABASE_ID,
     filter: {
@@ -71,6 +71,6 @@ exports.set = async function() {
   return
 }
 
-exports.expire = function() {
+exports.expire = function () {
   return fs.unlinkSync(BLOG_INDEX_CACHE)
 }
